fix(db): report real traffic totals and hand out invoice numbers safely
SUM over a bigint column returns numeric in Postgres, which reached the
repository as pgtype.Numeric behind an interface{}. The int64 type assertion
never matched and silently fell through to zero, so every user and every admin
card showed 0 bytes of traffic, always. The query now casts to bigint and the
generated types are int64.

Invoice numbers came from MAX(invoice_id)+1, so two checkouts running at the
same moment drew the same number and the loser failed on the unique
constraint — a broken payment for a user who did nothing wrong. They now come
from a sequence, seeded past the highest number already stored.

+++ internal/server/database/queries/history.sql
@@ -24,8 +24,8 @@ SELECT COUNT(*) FROM user_history WHERE user_id = $1;
 -- name: GetHistoryStats :one
 SELECT
     COUNT(*) AS total_connections,
-    COALESCE(SUM(bytes_sent), 0) AS total_bytes_sent,
-    COALESCE(SUM(bytes_received), 0) AS total_bytes_received
+    COALESCE(SUM(bytes_sent), 0)::bigint AS total_bytes_sent,
+    COALESCE(SUM(bytes_received), 0)::bigint AS total_bytes_received
 FROM user_history WHERE user_id = $1;
 
 -- name: DeleteHistoryOlderThan :execrows
